Split out ElevenLabs functions
